Moved image processing to estimator, used gui.cam in estimator
